requirements.txt
Django>=3.2.0,<4.2.0
djangorestframework>=3.12.0,<3.15.0

flake8>=4.0.0,<6.1.0
django-filter==23.3
factory-boy==3.3.0
Faker==15.3.4
freezegun==1.2.2
drf-yasg==1.21.7
